requirements.txt
altgraph==0.17.3
amqp==5.1.1
apturl==0.5.2
asgiref==3.5.2
asttokens==2.1.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.27.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
bottle==0.12.23
bottle-websocket==0.2.9
Brlapi==0.8.3
celery==5.2.7
certifi==2022.12.7
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
comm==0.1.1
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
DateTime==4.7
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.0.6
django-celery-beat==2.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
duplicity==0.8.21
Eel==0.14.0
entrypoints==0.4
evdev==1.4.0
exceptiongroup==1.1.0
executing==1.2.0
fake-useragent==1.1.1
fasteners==0.14.1
filelock==3.8.0
Flask==2.0.1
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
gyp==0.1
h11==0.14.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.6.4
ipykernel==6.18.1
ipython==8.7.0
itsdangerous==2.1.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.0.3
jupyter_client==7.4.7
jupyter_core==5.1.0
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
maticalgos==0.70
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.5.6
netifaces==0.11.0
nsepy==0.8
nsepython==0.0.973
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandas-ta==0.3.14b0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.33
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
PyChromecast==9.4.0
pycryptodomex==3.11.0
pycups==2.0.1
pydbus==0.6.0
Pygments==2.13.0
PyGObject==3.42.1
pyinotify==0.9.6
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.1
python-xlib==0.29
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
redis==4.5.1
reportlab==3.6.8
requests==2.25.1
scipy==1.9.3
SecretStorage==3.3.1
selenium==4.8.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.2
system76driver==20.4.74
systemd-python==234
TA-Lib==0.4.25
tk==0.1.0
tkintertable==1.3.3
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.16.5
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==3.8.5
websocket==0.2.1
websocket-client==1.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zeroconf==0.38.3
zipp==1.0.0
zope.event==4.6
zope.interface==5.5.2
